backend/requirements.txt
pydantic
python-dotenv
openai
argparse
numpy
os
argparse
json
supabase
anthropic